Check for env.rb via ENV_LOADED instead of respond_to?(:require)
The rubygem 'polyglot', which is depended on by Cucumber via Treetop,
redefines the Kernel#require method, making it public instead of
private. This is pretty horrible. Libraries should not break the
behaviour of the stdlib. :(

This commit implements a workaround for that particular issue, with
a test that does the same thing polyglot does.

Closes #287